begin to style the create draffle page
