+++ app/javascript/packs/application.scss
@@ -20,5 +20,6 @@
 @import "../stylesheets/collections/bulk-edit";
 @import "../stylesheets/objectives/objectives-manager";
 @import "../stylesheets/form/edit_user.scss";
+@import "../stylesheets/form/form_basic.scss";
 @import "../stylesheets/draffle/draffle-edit.scss";
 @import "../stylesheets/draffle/draffle-table.scss";
