+++ app/javascript/stylesheets/draffle/draffle-edit.scss
@@ -1,20 +1,24 @@
 .draffle-edit {
   &__print-selector {
-    width: 375px;
-    height: 300px;
+    width: 100%;
     display: flex;
     gap: 5%;
-    transition: margin-left 0.3s;
+
+    &.animate {
+      transition: margin-left 0.3s;
+    }
 
     &--print-option {
       width: 30%;
       flex: 1 0 auto;
       display: flex;
       align-items: center;
-      transition: width 0.3s, filter 0.3s;
       filter: grayscale(100%);
       position: relative;
-
+      
+      &.animate {
+        transition: width 0.3s, filter 0.3s;
+      }
 
       &.active {
         width: 50%;
