+++ app/javascript/stylesheets/form/edit_user.scss
@@ -61,78 +61,4 @@ form.edit_user {
       background: #7289DA;
     }
   }
-
-  &__fields {
-    display: flex;
-    flex-direction: column;
-    background: white;
-    border-radius: 0.5rem;
-    padding: 0.5rem 1rem 1.25rem;
-    margin: 1rem 0;
-    width: 100%;
-  }
-
-  &__text-field {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-bottom: 3px solid $color-highlight-mid-import;
-    transition: border 0.3s;
-
-    label {
-      font-weight: 600;
-      font-size: 1.2rem;
-      transition: font-size 0.3s, color 0.3s, font-weight 0.3s;
-    }
-
-    input[type=text]:not(.browser-default),
-    input[type=email]:not(.browser-default) {
-      margin: 0;
-      border: 0;
-      text-align: right;
-      font-size: 1rem;
-      color: #989898;
-      transition: font-size 0.3s, color 0.3s, font-weight 0.3s;
-
-      &:focus {
-        font-size: 1.2rem;
-        color: $color-nuetral;
-        font-weight: 600;
-        box-shadow: none;
-        border-bottom: 0;
-      }
-    }
-
-    &:focus-within {
-      label {
-        font-size: 1rem;
-        font-weight: 400;
-        color: #989898;
-      }
-    }
-
-    &:focus-within, &:hover {
-      border-color: lighten($color-accent, 5%);
-    }
-  }
-
-  &__switch {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    margin: 1rem 0;
-    padding: 0 1rem;
-
-    label {
-      font-size: 1.2rem;
-      color: $color-nuetral;
-      font-weight: 600;
-    }
-
-    div.switch label {
-      font-size: 1rem;
-      color: #989898;
-    }
-  }
 }
