+++ app/javascript/stylesheets/form/form_basic.scss
@@ -1,0 +1,89 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  background: white;
+  border-radius: 0.5rem;
+  padding: 0.5rem 1rem 1.25rem;
+  margin: 1rem 0;
+  width: 100%;
+
+  &__field {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    border-bottom: 3px solid $color-highlight-mid-import;
+    transition: border 0.3s;
+  
+    label {
+      font-weight: 600;
+      font-size: 1.2rem;
+      transition: font-size 0.3s, color 0.3s, font-weight 0.3s;
+    }
+
+    
+  
+    input[type=text]:not(.browser-default),
+    input[type=email]:not(.browser-default),
+    input[type=number]:not(.browser-default),
+    textarea {
+      margin: 0;
+      border: 0;
+      flex: 1 1 100%;
+      margin-left: 0;
+      box-shadow: none;
+      text-align: right;
+      font-size: 1rem;
+      color: #989898;
+      transition: font-size 0.3s, color 0.3s, font-weight 0.3s;
+  
+      &:focus {
+        font-size: 1.2rem;
+        color: $color-nuetral;
+        font-weight: 600;
+        box-shadow: none;
+        border-bottom: 0;
+        outline: 0 solid transparent;
+      }
+    }
+  
+    &:focus-within {
+      label {
+        font-size: 1rem;
+        font-weight: 400;
+        color: #989898;
+      }
+    }
+  
+    &:focus-within, &:hover {
+      border-color: lighten($color-accent, 5%);
+    }
+  }
+
+  textarea {
+    height: 9rem;
+    &:focus {
+      font-size: 1rem;
+    }
+  }
+  
+  &__switch {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    margin: 1rem 0;
+    padding: 0 1rem;
+  
+    label {
+      font-size: 1.2rem;
+      color: $color-nuetral;
+      font-weight: 600;
+    }
+  
+    div.switch label {
+      font-size: 1rem;
+      color: #989898;
+    }
+  }
+}
+
