+++ app/views/users/_user_fields.html.erb
@@ -17,10 +17,10 @@
     <% end %>
   <% end %>
 
-  <div class="dashboard_edit-user-details__fields">
-    <%= f.input :name, wrapper_html: { class: 'dashboard_edit-user-details__text-field' } %>
-    <%= f.input :email, wrapper_html: { class: 'dashboard_edit-user-details__text-field'} %>
-    <%= f.input :pronouns, wrapper_html: { class: 'dashboard_edit-user-details__text-field'} %>
+  <div class="form">
+    <%= f.input :name, wrapper_html: { class: 'form__field' } %>
+    <%= f.input :email, wrapper_html: { class: 'form__field'} %>
+    <%= f.input :pronouns, wrapper_html: { class: 'form__field'} %>
   </div>
   <div style="margin: 1rem 0;">
       <%= f.button :submit, class: "dashboard_edit-user-details_actions-submit__btn" %>
