+++ db/migrate/20231012183735_add_welcome_to_draffles.rb
@@ -1,0 +1,5 @@
+class AddWelcomeToDraffles < ActiveRecord::Migration[6.1]
+  def change
+    add_column :draffles, :welcome, :text, default: "Welcome to a New Draffle!"
+  end
+end
