+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2023_10_10_173747) do
+ActiveRecord::Schema.define(version: 2023_10_12_183735) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -83,6 +83,7 @@ ActiveRecord::Schema.define(version: 2023_10_10_173747) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.bigint "discord_thread_id"
+    t.text "welcome", default: "Welcome to a New Draffle!"
   end
 
   create_table "exchanges", force: :cascade do |t|
